requirements.txt
ib_insync==0.9.70
gspread==5.7.1
oauth2client==4.1.3
python-telegram-bot==13.7
pandas==1.5.3